Create Authorizations context and move resources to that
