+++ lib/id_werk/services/services.ex
@@ -199,107 +199,4 @@ defmodule IdWerk.Services do
   def change_scope(%Scope{} = scope) do
     Scope.changeset(scope, %{})
   end
-
-  alias IdWerk.Services.Resource
-
-  @doc """
-  Returns the list of resources.
-
-  ## Examples
-
-      iex> list_resources()
-      [%Resource{}, ...]
-
-  """
-  def list_resources do
-    Repo.all(Resource)
-  end
-
-  @doc """
-  Gets a single resource.
-
-  Raises `Ecto.NoResultsError` if the Resource does not exist.
-
-  ## Examples
-
-      iex> get_resource!(123)
-      %Resource{}
-
-      iex> get_resource!(456)
-      ** (Ecto.NoResultsError)
-
-  """
-  def get_resource!(id), do: Repo.get!(Resource, id)
-
-  @doc """
-  Creates a resource.
-
-  ## Examples
-
-      iex> create_resource(%{field: value})
-      {:ok, %Resource{}}
-
-      iex> create_resource(%{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def create_resource(%{user: user} = attrs) do
-    attrs
-    |> Map.put(:group, user.group)
-    |> Map.delete(:user)
-    |> create_resource()
-  end
-
-  def create_resource(%{} = attrs) do
-    %Resource{}
-    |> Resource.create_changeset(attrs)
-    |> Repo.insert()
-  end
-
-  @doc """
-  Updates a resource.
-
-  ## Examples
-
-      iex> update_resource(resource, %{field: new_value})
-      {:ok, %Resource{}}
-
-      iex> update_resource(resource, %{field: bad_value})
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def update_resource(%Resource{} = resource, attrs) do
-    resource
-    |> Resource.changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Deletes a Resource.
-
-  ## Examples
-
-      iex> delete_resource(resource)
-      {:ok, %Resource{}}
-
-      iex> delete_resource(resource)
-      {:error, %Ecto.Changeset{}}
-
-  """
-  def delete_resource(%Resource{} = resource) do
-    Repo.delete(resource)
-  end
-
-  @doc """
-  Returns an `%Ecto.Changeset{}` for tracking resource changes.
-
-  ## Examples
-
-      iex> change_resource(resource)
-      %Ecto.Changeset{source: %Resource{}}
-
-  """
-  def change_resource(%Resource{} = resource) do
-    Resource.changeset(resource, %{})
-  end
 end
